bots/sudolink/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "sudolink"
version = "0.1.0"
description = "SudoLink Telegram bot – link expander, not a fact checker."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=21.3",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pydantic>=2.6",
    "openai>=1.40.0",
]
authors = [
    {name = "SudoLink Team"}
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3"
]

[tool.setuptools.packages.find]
where = ["."]
